Rename 강사님 피드백 후 패키지 수정
