+++ smilekarina/src/main/java/com/example/smilekarina/gift/domain/Gift.java
@@ -1,4 +1,4 @@
-package com.example.smilekarina.point.domain.gift;
+package com.example.smilekarina.gift.domain;
 
 // 선물
 import com.example.smilekarina.global.domain.BaseEntity;
