+++ smilekarina/src/main/java/com/example/smilekarina/gift/domain/GiftType.java
@@ -1,6 +1,6 @@
-package com.example.smilekarina.point.domain.gift;
+package com.example.smilekarina.gift.domain;
 
-import com.example.smilekarina.point.domain.CodeValue;
+import com.example.smilekarina.global.domain.CodeValue;
 import lombok.Getter;
 
 @Getter
