+++ smilekarina/src/main/java/com/example/smilekarina/point/domain/PointType.java
@@ -1,9 +1,10 @@
 package com.example.smilekarina.point.domain;
 
+import com.example.smilekarina.global.domain.CodeValue;
 import lombok.Getter;
 
 @Getter
-public enum PointType implements CodeValue{
+public enum PointType implements CodeValue {
     EVENT("E", "이벤트"), GIFT("G", "선물"), COUPON("C", "쿠폰"), ETC("T", "기타");
 
     private final String code;
