+++ smilekarina/src/main/java/com/example/smilekarina/receipt/domain/ReceiptAccept.java
@@ -1,4 +1,4 @@
-package com.example.smilekarina.point.domain.receipt;
+package com.example.smilekarina.receipt.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,13 +6,13 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-// 스마트 영수증 조회가 가능한 곳에서 적립한 포인트
+// 스마트 영수증 조회가 불가능한 곳에서 적립한 포인트
 @Builder
 @Entity
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class SmartReceipt {
+public class ReceiptAccept {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -23,9 +23,6 @@ public class SmartReceipt {
 
     @Column(nullable = false, name = "user_id")
     private Long userId;
-
-    @Column(nullable = false, name = "receipt_id")
-    private Long receiptId;
 
     @Column(nullable = false, name = "franchise_id")
     private Integer franchiseId;
